cdn/sync_service/requirements.txt
alembic==1.10.3
asyncpg==0.27.0
backoff==2.2.1
black==23.3.0
fastapi==0.95.0
gunicorn==20.1.0
httpx==0.23.3
isort==5.12.0
orjson==3.8.9
pydantic==1.10.7
pytest==7.2.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
SQLAlchemy==2.0.9
uvicorn==0.21.1
psycopg2-binary==2.9.6
